starting rework
- reorganization of functions into differents files for easy maintenance.
